dependency update and move prisma to services
