+++ package.json
@@ -3,13 +3,13 @@
     "version": "1.0.1",
     "private": true,
     "dependencies": {
-        "@material-ui/core": "^4.11.4",
-        "@prisma/client": "^2.23.0",
-        "next": "^10.2.1",
+        "@material-ui/core": "^4.12.3",
+        "@prisma/client": "^2.29.0",
+        "next": "^11.1.0",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
-        "react-hook-form": "^7.6.2",
-        "swr": "^0.5.5"
+        "react-hook-form": "^7.12.2",
+        "swr": "^0.5.6"
     },
     "scripts": {
         "build": "next build",
@@ -29,31 +29,31 @@
         "prisma": "prisma format && npm run migrate && npm run seed"
     },
     "devDependencies": {
-        "@axe-core/react": "^4.2.0",
-        "@babel/core": "^7.13.16",
-        "@babel/node": "^7.13.13",
-        "@babel/plugin-transform-typescript": "^7.13.0",
-        "@babel/preset-env": "^7.13.15",
-        "@babel/preset-typescript": "^7.13.0",
-        "@testing-library/dom": "^7.31.0",
-        "@testing-library/jest-dom": "^5.12.0",
-        "@testing-library/react": "^11.2.7",
-        "@types/node": "^15.3.1",
-        "@types/react": "^17.0.6",
-        "@types/react-dom": "^17.0.5",
-        "@typescript-eslint/eslint-plugin": "^4.24.0",
-        "@typescript-eslint/parser": "^4.24.0",
+        "@axe-core/react": "^4.2.2",
+        "@babel/core": "^7.15.0",
+        "@babel/node": "^7.14.9",
+        "@babel/plugin-transform-typescript": "^7.15.0",
+        "@babel/preset-env": "^7.15.0",
+        "@babel/preset-typescript": "^7.15.0",
+        "@testing-library/dom": "^8.1.0",
+        "@testing-library/jest-dom": "^5.14.1",
+        "@testing-library/react": "^12.0.0",
+        "@types/node": "^16.6.0",
+        "@types/react": "^17.0.17",
+        "@types/react-dom": "^17.0.9",
+        "@typescript-eslint/eslint-plugin": "^4.29.1",
+        "@typescript-eslint/parser": "^4.29.1",
         "babel-eslint": "^10.1.0",
         "babel-jest": "^26.6.3",
-        "eslint": "^7.26.0",
+        "eslint": "^7.32.0",
         "eslint-config-prettier": "^8.3.0",
         "eslint-plugin-prettier": "^3.4.0",
-        "eslint-plugin-react": "^7.23.2",
+        "eslint-plugin-react": "^7.24.0",
         "identity-obj-proxy": "^3.0.0",
         "jest": "^26.6.3",
-        "prettier": "^2.3.0",
-        "prisma": "^2.23.0",
-        "ts-node": "^10.0.0",
-        "typescript": "^4.2.4"
+        "prettier": "^2.3.2",
+        "prisma": "^2.29.0",
+        "ts-node": "^10.2.0",
+        "typescript": "^4.3.5"
     }
 }
