+++ pages/api/tool/index.ts
@@ -1,7 +1,7 @@
 // https://nextjs.org/docs/api-routes/introduction
 import { NextApiRequest, NextApiResponse } from 'next';
 import { ToolForm } from '../../../components/dialog/ToolDialog';
-import prisma from '../../../prisma/prisma';
+import prisma from '../../../services/prisma';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
     const { method } = req;
